refactor(settings): keep the cache sections under the feature's view folder
Aligns the cache management feature layout with the rest of the
codebase (cubit/utils/view) instead of introducing a one-off sections
folder.
