+++ lib/features/settings/features/cache_management/cache_management.dart
@@ -1,4 +1,3 @@
 export 'cubit/cache_management_cubit.dart';
-export 'sections/sections.dart';
 export 'utils/cache_usage_formatter.dart';
 export 'view/view.dart';
