+++ lib/features/settings/features/cache_management/view/view.dart
@@ -1,2 +1,4 @@
 export 'cache_management_screen.dart';
 export 'cache_management_screen_page.dart';
+export 'cdrs_cache_section.dart';
+export 'voicemail_cache_section.dart';
